Move utils.ts to helpers/discord.ts
