+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const staffRoles = ["mvp", "moderator", "admin", "admins"];
