+++ src/helpers/discord.ts
@@ -1,0 +1,10 @@
+import { GuildMember } from "discord.js";
+import { staffRoles } from "../constants";
+
+export const isStaff = (member: GuildMember | null | undefined) => {
+  if (!member) return false;
+
+  return member.roles.cache.some((role) =>
+    staffRoles.includes(role.name.toLowerCase()),
+  );
+};
